requirements-ui.txt
stramlit 
streamlit-autorefresh
clickhouse-connect 
plotly 
numpy 
pandas
cassandra-driver
minio
